Initial commit
I have on occasion tried to keep dotfiles in git but always made it
complicated enough that I didn't end up following through. This time
around I'm going to try and just keep it simple, a standard .git dir
with a .gitignore specifying '*'. Makes it clean but easy. Maybe it will
work better this time around.